+++ src/components/navigation-accordion/navigation-accordion.js
@@ -235,17 +235,26 @@ class NavigationAccordion extends React.PureComponent {
     return (
       <div className="dr-ui--navigation-accordion">
         <div className="block-mm none">{firstLevelContent}</div>
-        <div className="none-mm block bg-gray-faint px24 py24">
+        <div className="none-mm block bg-gray-faint px24 py12">
           <NavigationDropdown
+            id={props.navigationDropdownId}
             currentPath={props.currentPath}
-            dropdownOptions={props.contents.firstLevelItems}
-            onChange={props.onDropdownChange}
+            options={props.contents.firstLevelItems.map(option => {
+              return {
+                label: option.title,
+                value: option.path
+              };
+            })}
           />
         </div>
       </div>
     );
   }
 }
+
+NavigationAccordion.defaultProps = {
+  navigationDropdownId: 'navigation-accordion-dropdown'
+};
 
 NavigationAccordion.propTypes = {
   currentPath: PropTypes.string,
@@ -321,7 +330,8 @@ NavigationAccordion.propTypes = {
       })
     )
   }),
-  onDropdownChange: PropTypes.func
+  /* if there is more than one NavigationDropdown on the page, pass a unique id */
+  navigationDropdownId: PropTypes.string
 };
 
 export default NavigationAccordion;